backend/requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
audioop-lts==0.2.1
audioread==3.1.0
bcrypt==5.0.0
beautifulsoup4==4.13.3
blinker==1.9.0
bottle==0.13.2
bottle-websocket==0.2.9
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comtypes==1.4.10
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
Eel==0.18.1
fastapi==0.135.1
filelock==3.24.3
Flask==3.1.0
fonttools==4.61.1
fsspec==2026.2.0
future==1.0.0
gevent==24.11.1
gevent-websocket==0.10.1
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.1.1
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httpx==0.28.1
hugchat==0.4.18
huggingface_hub==1.4.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
more-itertools==10.8.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.2
narwhals==2.17.0
networkx==3.6.1
nltk==3.9.2
numba==0.64.0
numpy==2.4.2
openai-whisper==20250625
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
packaging==26.0
pandas==2.3.3
passlib==1.7.4
pathlib==1.0.1
pillow==12.1.1
platformdirs==4.9.2
pluggy==1.6.0
pooch==1.9.0
protobuf==6.33.5
pvporcupine==3.0.5
pyarrow==23.0.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.11.0
PyMsgBox==1.0.9
pyparsing==3.3.2
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyttsx3==2.98
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
pywin32==309
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==76.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sounddevice==0.5.5
soundfile==0.13.1
soupsieve==2.6
soxr==1.0.0
SpeechRecognition==3.14.1
SQLAlchemy==2.0.48
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.52.1
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
watchdog==6.0.0
Werkzeug==3.1.3
wikipedia==1.4.0
zope.event==5.0
zope.interface==7.2
